+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Version 67
 *in development*
 
 * 🔧 On Android 8 and newer, use system settings to configure notification settings.
+* 🔧 On Android 13 and newer, ask to allow notifications.
 
 #### 67.0.0 🧪
 *2023-01-27*